requirements.txt
# Core dependencies for Snowflake MCP Server
fastmcp>=2.9.2
snowflake-connector-python>=3.0.0
pydantic>=2.0.0
httpx>=0.28.1
cachetools>=5.0.0
sqlparse>=0.4.0
starlette>=0.47.1
uvicorn>=0.34.3
